requirements.txt
# S3verless Core Dependencies
# Generated from pyproject.toml - Use for pip installation

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Data Validation
pydantic[email]>=2.9.0
pydantic-settings>=2.6.0

# AWS S3
aiobotocore>=2.15.0
botocore>=1.35.0
boto3>=1.35.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Utilities
python-multipart>=0.0.12
click>=8.1.7

# Optional: For local development
# localstack>=3.0.0

